+++ src/internet/model/ipv4-global-routing.cc
@@ -192,9 +192,13 @@ Ipv4GlobalRouting::LookupGlobal(Ipv4Address dest, Ptr<NetDevice> oif)
     if (allRoutes.empty()) // if no host route is found
     {
         NS_LOG_LOGIC("Number of m_networkRoutes" << m_networkRoutes.size());
+        // store the length of the longest mask.
+        uint16_t longest_mask = 0;
         for (auto j = m_networkRoutes.begin(); j != m_networkRoutes.end(); j++)
         {
             Ipv4Mask mask = (*j)->GetDestNetworkMask();
+            uint16_t masklen = mask.GetPrefixLength();
+
             Ipv4Address entry = (*j)->GetDestNetwork();
             if (mask.IsMatch(dest, entry))
             {
@@ -206,8 +210,23 @@ Ipv4GlobalRouting::LookupGlobal(Ipv4Address dest, Ptr<NetDevice> oif)
                         continue;
                     }
                 }
-                allRoutes.push_back(*j);
                 NS_LOG_LOGIC(allRoutes.size() << "Found global network route" << *j);
+                if (masklen < longest_mask) // Not interested if got shorter mask
+                {
+                    NS_LOG_LOGIC("Previous match longer, skipping");
+                    continue;
+                }
+                else if (masklen == longest_mask)
+                {
+                    NS_LOG_LOGIC("Equal mask length, adding this to the list");
+                    allRoutes.push_back(*j);
+                }
+                else
+                {
+                    NS_LOG_LOGIC("Longer mask length found, clearing the list and adding");
+                    allRoutes.clear();
+                    allRoutes.push_back(*j);
+                }
             }
         }
     }